add client service

+++ src/main/java/ropandi/oauth/service/impl/ClientDetailsServiceImpl.java
@@ -28,8 +28,8 @@ public class ClientDetailsServiceImpl implements ClientDetailsService {
         String authorities = client.getAuthorities().stream().collect(Collectors.joining(","));
 		BaseClientDetails base =  new BaseClientDetails(client.getClientId(), resourceIds, scopes, grantTypes, authorities);
         base.setClientSecret(client.getClientSecret());
-        base.setAccessTokenValiditySeconds(client.getAccessTokenValiditySeconds());
-        base.setRefreshTokenValiditySeconds(client.getRefreshTokenValiditySeconds());
+        base.setAccessTokenValiditySeconds(client.getAccessTokenValiditySeconds().intValue());
+        base.setRefreshTokenValiditySeconds(client.getRefreshTokenValiditySeconds().intValue());
         base.setAdditionalInformation(client.getAdditionalInformations());
         base.setAutoApproveScopes(client.getScopes());
         return base;
